requirements.txt
# Core Dependencies
requests>=2.31.0
PyYAML>=6.0
mysql-connector-python>=8.0.33

# Shopify Dependencies
shopify-python-api>=1.0.1

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
python-dotenv>=1.0.0
click>=8.1.0
rich>=13.0.0

# Development Dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0

# Logging and Monitoring
structlog>=23.0.0
